style: notice 모듈을 notices 모듈로 통합
