+++ src/frontend/src/store/modules/notices.js
@@ -2,11 +2,15 @@ import { getAction } from "../../api";
 
 export default {
   state: {
-    notices: []
+    notices: [],
+    notice: []
   },
   mutations: {
     SET_NOTICES(state, data) {
       state.notices = data;
+    },
+    SET_NOTICE(state, data) {
+      state.notice = data;
     }
   },
   actions: {
@@ -17,11 +21,22 @@ export default {
       } catch (error) {
         console.log(error);
       }
+    },
+    async FETCH_NOTICE({ commit }, noticeId) {
+      try {
+        const { data } = await getAction(`/api/notices/${noticeId}`);
+        commit("SET_NOTICE", data);
+      } catch (error) {
+        console.log(error);
+      }
     }
   },
   getters: {
     fetchedNotices(state) {
       return state.notices;
+    },
+    fetchedNotice(state) {
+      return state.notice;
     }
   }
 };
